requirements.txt
absl-py==2.4.0
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1763409923949/work
attrs==25.4.0
catboost==1.2.10
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
cloudpickle==3.1.2
colorlog==6.10.1
comm @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_comm_1753453984/work
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.4.1
cycler==0.12.1
datasets==4.6.1
debugpy @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_debugpy_1769744978/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
dill==0.4.0
einops==0.8.2
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1756729339227/work
filelock==3.25.0
fire==0.7.1
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
graphviz==0.21
greenlet==3.3.2
grpcio==1.78.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.6.0
idna==3.11
ImageIO==2.37.3
imbalanced-learn==0.14.1
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ipykernel_1770566133/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ipython_1772790149/work
ipython_pygments_lexers @ file:///home/conda/feedstock_root/build_artifacts/ipython_pygments_lexers_1737123620466/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.6
joblib==1.5.3
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jupyter_client_1767954655/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jupyter_core_1760643864/work
kiwisolver==1.4.9
lazy-loader==0.5
lightgbm==4.6.0
lightning==2.6.1
lightning-utilities==0.15.3
llvmlite==0.46.0
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1761214490209/work
mdurl==0.1.2
medmnist==3.0.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
narwhals==2.17.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.6.1
numba==0.64.0
numpy==2.4.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
optuna==4.7.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1769093650/work
pandas==3.0.1
parso @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_parso_1770676243/work
patsy==1.0.2
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pillow==12.1.1
platformdirs==4.9.4
plotly==6.6.0
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1756321756983/work
propcache==0.4.1
protobuf==6.33.5
psutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_psutil_1769678155/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
pyparsing==3.3.2
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
pytorch-lightning==2.6.1
PyYAML==6.0.3
pyzmq @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pyzmq_1771716961/work
regex==2026.2.28
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
sentry-sdk==2.54.0
setuptools==82.0.0
shap==0.51.0
shellingham==1.5.4
simsimd==6.5.16
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
sklearn-compat==0.1.5
slicer==0.0.8
smmap==5.0.2
SQLAlchemy==2.0.48
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
statsmodels==0.14.6
stringzilla==4.6.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.3.0
threadpoolctl==3.6.0
tifffile==2026.3.3
timm==1.0.25
tokenizers==0.22.2
torch==2.10.0
torchmetrics==1.8.2
torchvision==0.25.0
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1765836395280/work
tqdm==4.67.3
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
transformers==5.3.0
triton==3.6.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wandb==0.25.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1770634525285/work
Werkzeug==3.1.6
wheel==0.46.3
xgboost==3.2.0
xxhash==3.6.0
yarl==1.23.0

